move recLogo.svg to new assets folder
