+++ src/ui/app.js
@@ -16,7 +16,7 @@ import "./recordembed";
 
 import { BtrixClient } from "./upload";
 
-import wrRec from "../../assets/recLogo.svg";
+import wrRec from "../assets/icons/recLogo.svg";
 import awpLogo from "../assets/brand/archivewebpage-icon-color.svg";
 import awpBrandLockupColor from "../assets/brand/archivewebpage-lockup-color.svg";
 import prettyBytes from "pretty-bytes";
